pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jania"
version = "0.3.0"
description = "Multipurpose python utils"
readme = "README.md"
license = {file = "LICENSE"}
authors = [{name = "Julian Ania", email = "julian.ania@zebraventures.eu"}]
keywords = ["utils", "jania"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "toml",
    "PyYAML",
    "python-dotenv",
    "pdf2image",
    "Pillow",        # Muy recomendable añadirlo explícitamente
    "openai"         # Si la parte de LLM es fundamental para tu librería
]

[project.urls]
"Homepage" = "https://github.com/julianin/JANIALIB"
